+++ iso20022-core/src/main/java/com/prowidesoftware/swift/model/MxBusinessProcess.java
@@ -15,10 +15,14 @@
  */
 package com.prowidesoftware.swift.model;
 
+import com.prowidesoftware.deprecation.ProwideDeprecated;
+import com.prowidesoftware.deprecation.TargetYear;
+
+import java.util.Collection;
+
 /**
  * SWIFT business process classification for MX messages.
  *
- * @author www.prowidesoftware.com
  * @since 7.0
  */
 public enum MxBusinessProcess {
@@ -26,8 +30,12 @@ public enum MxBusinessProcess {
     admi("Administration"),
     auth("Authorities"),
     caaa("Acceptor to Acquirer Card Transactions"),
+    caad("Card Administration"),
     caam("ATM Management"),
+    cafc("Fee Collection"),
+    cain("Acquirer to Issuer Card Transactions"),
     camt("Cash Management"),
+    casp("Sale to POI Card Transactions"),
     catm("Terminal Management"),
     catp("ATM Card Transactions"),
     cbrf(""),
@@ -47,7 +55,14 @@ public enum MxBusinessProcess {
     setr("Securities Trade"),
     supl("Supplementary Data"),
     trck("Payments Tracker"),
+
+    /**
+     * @deprecated Obsolete to be withdrawn (non-ISO)
+     */
+    @ProwideDeprecated(phase2 = TargetYear.SRU2023)
+    @Deprecated
     trea("Treasury"),
+
     tsin("Trade Services Initiation"),
     tsmt("Trade Services Management"),
     tsrv("Trade Services"),